money_transactions table decimal amount field type

+++ database/migrations/2019_02_03_183030_create_money_transactions_table.php
@@ -17,7 +17,7 @@ class CreateMoneyTransactionsTable extends Migration
             $table->increments('id');
             $table->unsignedInteger('user_id');
             $table->string('status');
-            $table->double('amount');
+            $table->decimal('amount');
             $table->string('currency');
             $table->string('card_number')->nullable();
             $table->text('info')->nullable();
